gave ButtonState its own file
